Rename instances of Users... to User... in exercise 2
