+++ exercises/02-react-stateless/src/index.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import ReactDOM from 'react-dom'
 
 import { Root } from 'modules/root/components/root'
-import { User, AddUserFunc } from 'modules/users/users-types'
+import { User, AddUserFunc } from 'modules/users/user-types'
 
 interface State {
   title: string
